get_windows_pass/requirements.txt
appdirs==1.4.3
-e git+https://github.com/c2devel/boto.git@2.12.0-CROC8#egg=boto
packaging==16.8
pycrypto==2.6.1
pyparsing==2.2.0
six==1.10.0
